learned something about lat long, switched db to decimal, created extral localized data

+++ models/message.js
@@ -1,7 +1,7 @@
 module.exports = function(sequelize, DataTypes) {
   return sequelize.define("Message", {
-    latitude: DataTypes.FLOAT,
-    longitude: DataTypes.FLOAT,
+    latitude: DataTypes.DECIMAL(10,7),
+    longitude: DataTypes.DECIMAL(10,7),
     username: DataTypes.STRING,
     message: DataTypes.TEXT
   });
